services/correlation-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
nats-py==2.6.0
python-dateutil==2.8.2
